medicine_verification/requirements.txt
# Medicine Verification System - Python Dependencies

# Web Framework
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# OCR Libraries (choose one or both)
# Option 1: Tesseract OCR
pytesseract==0.3.10
Pillow==10.1.0

# Option 2: EasyOCR (alternative, more accurate but larger)
easyocr==1.7.0

# Image Processing (optional, for preprocessing)
opencv-python==4.8.1.78
numpy==1.24.3

# Email Notifications
# Note: email libraries are usually included with Python, but listed here for reference

# SMS Notifications (optional)
# twilio==8.10.0  # Uncomment if using Twilio for SMS

# Firebase Integration (optional)
# firebase-admin==6.4.0  # Uncomment if integrating with Firebase

# HTTP Requests (for downloading images from Firebase Storage URLs)
requests==2.31.0

# Utilities
python-dotenv==1.0.0  # For environment variables
